FIX: resolve __dirname to expand paths
Running on Linux this error comes up because `__dirname` is just `.` (current working directory). 

`Error: ENOENT: no such file or directory, scandir 'user/mods/./database/'`

This updates to use the `path` module to expand short names such as `.`, `~`, and so on.

I have made the change and tested and it works on Windows and Linux.